// packages/push-notification/src/push.notification.enums.ts
export enum FirebaseSendMessageStatus {
  INVALID_DATA = 'INVALID_DATA',
  LIMITED = 'LIMITED',
  INVALID_TOKEN = 'INVALID_TOKEN',
  EXCEEDED = 'EXCEEDED',
  TOO_MANY_TOPICS = 'TOO_MANY_TOPICS',
  AUTH = 'AUTH',
  CERDENTIALS = 'CERDENTIALS',
  UNAVAILABLE = 'UNAVAILABLE',
  INTERNAL = 'INTERNAL',
  UNKNOWN = 'UNKNOWN',
}